# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HSE Platform API
  description: "\nHSE platform API — Phase 0 Foundation (spec `docs/specs/0-foundation.md` v1.0).\n\n\
    * Auth: `POST /api/v1/auth/login` sets the httpOnly SameSite=Lax cookie `hse_session` (JWT) and\n\
    \  returns `{access_token, user}`. Send the cookie or `Authorization: Bearer <token>`.\n* Errors:\
    \ `{\"detail\": {\"code\", \"message\", \"message_ar\", \"errors\"}}`; 401 unauthenticated,\n  403\
    \ forbidden, 404 not found / out of scope, 409 conflict / invalid transition, 422 validation.\n* Lists:\
    \ `?page=1&page_size=50` → `{items, total, page, page_size}`.\n* Timestamps are UTC; display and day\
    \ boundaries use the project timezone (Asia/Riyadh).\n"
  version: 0.1.0
paths:
...
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Log in with email and password
      description: Sets the httpOnly `hse_session` cookie and returns the same JWT. Unknown email and
        wrong password give an identical 401 INVALID_CREDENTIALS. After 5 failures in 15 min the account
        is locked for 15 min (401 ACCOUNT_LOCKED).
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout:
    post:
      tags:
      - auth
      summary: Log out (revokes the session, clears the cookie)
      operationId: logout
      responses:
        '204':
          description: Successful Response
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - sessionCookie: []
      - bearerAuth: []
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: Current user with resolved roles and capabilities
      operationId: get_me
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - sessionCookie: []
      - bearerAuth: []
    patch:
      tags:
      - auth
      summary: Edit own profile (name, mobile, language)
      operationId: update_me
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - sessionCookie: []
      - bearerAuth: []
  /api/v1/auth/me/password:
    post:
      tags:
      - auth
      summary: Change own password
      description: Other sessions of the user are revoked.
      operationId: change_password
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
        required: true
      responses:
        '204':
          description: Successful Response
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - sessionCookie: []
      - bearerAuth: []
  /api/v1/auth/privacy-notice/ack:
    post:
      tags:
      - auth
      summary: Acknowledge the current privacy notice
      description: Allowed before acknowledgement (rule 6). 409 PRIVACY_NOTICE_VERSION_MISMATCH if `version`
        is not the current one.
      operationId: acknowledge_privacy_notice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrivacyAckRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - sessionCookie: []
      - bearerAuth: []
  /api/v1/auth/invitations/validate:
    post:
      tags:
      - auth
      summary: Check an invite token (no auth)
      description: 404 INVITE_INVALID for unknown/used/superseded tokens; 410 INVITE_EXPIRED after 72
        h.
      operationId: validate_invitation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationTokenRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationInfo'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Token expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/invitations/accept:
    post:
      tags:
      - auth
      summary: 'Accept an invite: set password, acknowledge privacy notice, log in'
      operationId: accept_invitation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationAcceptRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Token expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/password-reset:
    post:
      tags:
      - auth
      summary: Request a password reset email (always 202)
      operationId: request_password_reset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/password-reset/confirm:
    post:
      tags:
      - auth
      summary: Set a new password with a reset token (single use, 60 min)
      operationId: confirm_password_reset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
        required: true
      responses:
        '204':
          description: Successful Response
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Token expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    AcceptedResponse:
      properties:
        accepted:
          type: boolean
          const: true
          title: Accepted
          default: true
        message:
          type: string
          title: Message
          description: Generic message; never reveals whether the email exists.
      type: object
      required:
      - message
      title: AcceptedResponse
...
    Capability:
      type: string
      enum:
      - project.manage
      - project.view
      - site_zone.manage
      - site_zone.view
      - contractor.create
      - contractor.approve
      - engagement.manage
      - contractor.view
      - contractor.view_contacts
      - user.invite
      - user.view_directory
      - user.view_contacts
      - user.manage_status
      - settings.edit
      - settings.view
      - audit_log.read
      - history.view
      - export.lists
      - profile.edit_own
      title: Capability
      description: Rows of the permission matrix (§5.10), numbered 1-19 in spec order.
    CapabilityGrant:
      properties:
        capability:
          $ref: '#/components/schemas/Capability'
        scope:
          $ref: '#/components/schemas/CapabilityScope'
          description: How far the capability reaches on this project (matrix legend §5.10).
      type: object
      required:
      - capability
      - scope
      title: CapabilityGrant
    CapabilityScope:
      type: string
      enum:
      - all
      - project
      - sites
      - contractor_tree
      - own_engagement
      title: CapabilityScope
      description: 'Matrix legend (§5.10): all projects, assigned project, assigned sites, contractor
        tree

        (own engagement + downstream subcontractors), own engagement only.'
...
    EmployerType:
      type: string
      enum:
      - client
      - pmc_consultant
      - contractor
      title: EmployerType
...
    ErrorCode:
      type: string
      enum:
      - UNAUTHENTICATED
      - INVALID_CREDENTIALS
      - ACCOUNT_LOCKED
      - SESSION_EXPIRED
      - PRIVACY_ACK_REQUIRED
      - PRIVACY_NOTICE_VERSION_MISMATCH
      - INVITE_INVALID
      - INVITE_EXPIRED
      - RESET_TOKEN_INVALID
      - WEAK_PASSWORD
      - CURRENT_PASSWORD_INCORRECT
      - FORBIDDEN
      - READ_ONLY_ROLE
      - CONTRACTOR_SUSPENDED
      - SELF_MODIFICATION_FORBIDDEN
      - ROLE_NOT_ASSIGNABLE
      - NOT_FOUND
      - VALIDATION_ERROR
      - DUPLICATE_VALUE
      - INVALID_TRANSITION
      - TRANSITION_CONDITION_NOT_MET
      - PROJECT_CLOSED
      - SITE_INACTIVE
      - ZONE_ARCHIVED
      - CONTRACTOR_NOT_APPROVED
      - LAST_HSE_MANAGER
      - SOD_CONFLICT
      - NOT_IMPLEMENTED
      - INTERNAL_ERROR
      title: ErrorCode
      description: 'Machine-readable error codes. Stable: the frontend translates these.'
    ErrorDetail:
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
          description: Stable machine code; map it to EN/AR UI text.
        message:
          type: string
          title: Message
          description: English, human-readable message.
        message_ar:
          anyOf:
          - type: string
          - type: 'null'
          title: Message Ar
          description: Arabic message, when available.
        errors:
          anyOf:
          - items:
              $ref: '#/components/schemas/FieldError'
            type: array
          - type: 'null'
          title: Errors
          description: Field-level errors (422 validation, 409 duplicates).
      type: object
      required:
      - code
      - message
      title: ErrorDetail
    ErrorResponse:
      properties:
        detail:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - detail
      title: ErrorResponse
      description: Envelope for every 4xx/5xx response.
...
    FieldError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Loc
          description: Path to the field, e.g. ["body", "zone_type"].
        msg:
          type: string
          title: Msg
          description: English description of the problem.
        type:
          type: string
          title: Type
          description: Machine-readable error type, e.g. 'missing', 'value_error'.
      type: object
      required:
      - loc
      - msg
      - type
      title: FieldError
      description: One field-level problem (used with VALIDATION_ERROR and DUPLICATE_VALUE).
...
    InvitationAcceptRequest:
      properties:
        token:
          type: string
          maxLength: 200
          minLength: 16
          title: Token
        password:
          type: string
          maxLength: 256
          minLength: 12
          title: Password
          description: At least 12 characters, at least 3 of 4 classes (upper, lower, digit, symbol),
            not equal to the email local part and not a common breached password (spec §5.1 rule 2).
        privacy_notice_version:
          type: string
          title: Privacy Notice Version
          description: Acknowledged notice version.
        preferred_language:
          anyOf:
          - $ref: '#/components/schemas/Language'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - token
      - password
      - privacy_notice_version
      title: InvitationAcceptRequest
    InvitationInfo:
      properties:
        email:
          type: string
          title: Email
        full_name_en:
          type: string
          title: Full Name En
        full_name_ar:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name Ar
        preferred_language:
          $ref: '#/components/schemas/Language'
        expires_at:
          type: string
          format: date-time
          title: Expires At
        privacy_notice:
          $ref: '#/components/schemas/PrivacyNotice'
      type: object
      required:
      - email
      - full_name_en
      - full_name_ar
      - preferred_language
      - expires_at
      - privacy_notice
      title: InvitationInfo
    InvitationTokenRequest:
      properties:
        token:
          type: string
          maxLength: 200
          minLength: 16
          title: Token
          description: Token from the invite link.
      additionalProperties: false
      type: object
      required:
      - token
      title: InvitationTokenRequest
...
    Language:
      type: string
      enum:
      - en
      - ar
      title: Language
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
          description: Login email (case-insensitive).
        password:
          type: string
          maxLength: 256
          minLength: 1
          title: Password
      additionalProperties: false
      type: object
      required:
      - email
      - password
      title: LoginRequest
    LoginResponse:
      properties:
        access_token:
          type: string
          title: Access Token
          description: JWT; same value as the hse_session cookie.
        token_type:
          type: string
          const: bearer
          title: Token Type
          default: bearer
        expires_at:
          type: string
          format: date-time
          title: Expires At
          description: Absolute session expiry (12 h after login).
        idle_timeout_minutes:
          type: integer
          title: Idle Timeout Minutes
          description: Session ends after this many idle minutes.
        user:
          $ref: '#/components/schemas/Me'
      type: object
      required:
      - access_token
      - expires_at
      - idle_timeout_minutes
      - user
      title: LoginResponse
    Me:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          title: Email
        full_name_en:
          type: string
          title: Full Name En
        full_name_ar:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name Ar
        mobile:
          anyOf:
          - type: string
          - type: 'null'
          title: Mobile
        employer_type:
          $ref: '#/components/schemas/EmployerType'
        employer_contractor_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Employer Contractor Id
        job_title:
          anyOf:
          - type: string
          - type: 'null'
          title: Job Title
        preferred_language:
          $ref: '#/components/schemas/Language'
        display_language:
          $ref: '#/components/schemas/Language'
          description: 'Resolved UI language: preferred_language, then project default, then en (rule
            43).'
        status:
          $ref: '#/components/schemas/UserStatus'
        mfa_enabled:
          type: boolean
          title: Mfa Enabled
        last_login_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Login At
        privacy_notice_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Privacy Notice Version
          description: Version the user acknowledged.
        privacy_notice_ack_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Privacy Notice Ack At
        privacy_ack_required:
          type: boolean
          title: Privacy Ack Required
          description: True when the current notice version is not acknowledged. While true every API
            except privacy-notice ack and logout returns 403 PRIVACY_ACK_REQUIRED.
        is_hse_manager:
          type: boolean
          title: Is Hse Manager
        org_capabilities:
          items:
            $ref: '#/components/schemas/Capability'
          type: array
          title: Org Capabilities
          description: Organisation-wide capabilities (from hse_manager); empty for other roles.
        projects:
          items:
            $ref: '#/components/schemas/ProjectAccess'
          type: array
          title: Projects
        role_assignments:
          items:
            $ref: '#/components/schemas/RoleAssignmentRead'
          type: array
          title: Role Assignments
          description: Active assignments only.
      type: object
      required:
      - id
      - email
      - full_name_en
      - full_name_ar
      - mobile
      - employer_type
      - employer_contractor_id
      - job_title
      - preferred_language
      - display_language
      - status
      - mfa_enabled
      - last_login_at
      - privacy_notice_version
      - privacy_notice_ack_at
      - privacy_ack_required
      - is_hse_manager
      - org_capabilities
      - projects
      - role_assignments
      title: Me
      description: The authenticated user, with resolved permissions for UI gating (server still enforces).
    MeUpdate:
      properties:
        full_name_en:
          anyOf:
          - type: string
            maxLength: 120
            minLength: 1
          - type: 'null'
          title: Full Name En
        full_name_ar:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Full Name Ar
        mobile:
          anyOf:
          - type: string
            pattern: ^\+[1-9]\d{7,14}$
          - type: 'null'
          title: Mobile
          description: E.164.
        preferred_language:
          anyOf:
          - $ref: '#/components/schemas/Language'
          - type: 'null'
      additionalProperties: false
      type: object
      title: MeUpdate
      description: Own-profile edit (capability 19).
...
    PasswordChangeRequest:
      properties:
        current_password:
          type: string
          maxLength: 256
          minLength: 1
          title: Current Password
        new_password:
          type: string
          maxLength: 256
          minLength: 12
          title: New Password
          description: At least 12 characters, at least 3 of 4 classes (upper, lower, digit, symbol),
            not equal to the email local part and not a common breached password (spec §5.1 rule 2).
      additionalProperties: false
      type: object
      required:
      - current_password
      - new_password
      title: PasswordChangeRequest
    PasswordResetConfirm:
      properties:
        token:
          type: string
          maxLength: 200
          minLength: 16
          title: Token
        new_password:
          type: string
          maxLength: 256
          minLength: 12
          title: New Password
          description: At least 12 characters, at least 3 of 4 classes (upper, lower, digit, symbol),
            not equal to the email local part and not a common breached password (spec §5.1 rule 2).
      additionalProperties: false
      type: object
      required:
      - token
      - new_password
      title: PasswordResetConfirm
    PasswordResetRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
      additionalProperties: false
      type: object
      required:
      - email
      title: PasswordResetRequest
    PrivacyAckRequest:
      properties:
        version:
          type: string
          title: Version
          description: Must equal the current notice version.
      additionalProperties: false
      type: object
      required:
      - version
      title: PrivacyAckRequest
    PrivacyNotice:
      properties:
        version:
          type: string
          title: Version
          examples:
          - PN-1.0
        text_en:
          type: string
          title: Text En
        text_ar:
          type: string
          title: Text Ar
      type: object
      required:
      - version
      - text_en
      - text_ar
      title: PrivacyNotice
    ProjectAccess:
      properties:
        project_id:
          type: string
          format: uuid
          title: Project Id
        project_code:
          type: string
          title: Project Code
        roles:
          items:
            $ref: '#/components/schemas/Role'
          type: array
          title: Roles
        capabilities:
          items:
            $ref: '#/components/schemas/CapabilityGrant'
          type: array
          title: Capabilities
        site_ids:
          anyOf:
          - items:
              type: string
              format: uuid
            type: array
          - type: 'null'
          title: Site Ids
          description: Sites the user is restricted to; null = all sites of the project (rule 11).
        contractor_engagement_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Contractor Engagement Ids
          description: Engagements the user is scoped to (Contractor HSE Rep / Permit Receiver).
        read_only:
          type: boolean
          title: Read Only
          description: True when the user only holds viewer_client here.
      type: object
      required:
      - project_id
      - project_code
      - roles
      - capabilities
      - site_ids
      - contractor_engagement_ids
      - read_only
      title: ProjectAccess
      description: What the current user may do on one project (union of active assignments, rule 9).
...
    Role:
      type: string
      enum:
      - hse_manager
      - hse_officer
      - site_engineer
      - permit_issuer
      - permit_receiver
      - contractor_hse_rep
      - viewer_client
      title: Role
      description: Platform roles (§3.8). hse_manager is organisation-wide; all others are per project.
...
    RoleAssignmentRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        role:
          $ref: '#/components/schemas/Role'
        project_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Project Id
          description: Null only for hse_manager (org-wide).
        site_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Site Ids
          description: Empty = all sites of the project.
        contractor_engagement_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Contractor Engagement Id
          description: Required for contractor_hse_rep and permit_receiver.
        valid_from:
          type: string
          format: date
          title: Valid From
        valid_to:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Valid To
        is_active:
          type: boolean
          title: Is Active
          description: True if today (Asia/Riyadh) is within the validity.
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
        created_at:
          type: string
          format: date-time
          title: Created At
        created_by_user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Created By User Id
      type: object
      required:
      - id
      - user_id
      - role
      - project_id
      - site_ids
      - contractor_engagement_id
      - valid_from
      - valid_to
      - is_active
      - revoked_at
      - created_at
      - created_by_user_id
      title: RoleAssignmentRead
      description: A role held by a user, optionally scoped to a project, sites and a contractor.
...
    UserStatus:
      type: string
      enum:
      - invited
      - active
      - locked
      - deactivated
      title: UserStatus
      description: User lifecycle (§4.3).
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      description: httpOnly SameSite=Lax cookie set by POST /auth/login.
      in: cookie
      name: hse_session
    bearerAuth:
      type: http
      description: 'Same JWT as the cookie, as `Authorization: Bearer <token>`.'
      scheme: bearer